# perfs: optimize serialize_string_value