I have found `serialize_string_value` to be a bottleneck in my code, and this was fixed by hotpatching and replacing it to use str.replace instead of "".join.

Benchmark:

```
def slow_serialize_string_value(value):
    return ''.join(
        r'\"' if c == '"' else
        r'\\' if c == '\\' else
        r'\A ' if c == '\n' else
        r'\D ' if c == '\r' else
        r'\C ' if c == '\f' else
        c
        for c in value
    )

_replacement_string_value = {
        '"': r"\"",
        "\\": r"\\",
        "\n": r"\A ",
        "\r": r"\D ",
        "\f": r"\C ",
    }
_re_string_value = "".join(re.escape(e) for e in _replacement_string_value.keys())
_re_string_value = re.compile("["+ _re_string_value + "]", re.MULTILINE )
def _serialize_string_value_match(match):
    return _replacement_string_value[match.group(0)]
def fast_serialize_string_value(value):
    return _re_string_value.sub(_serialize_string_value_match, value)

string = "".join(chr(random.randint(0, 255)) for _ in range(10_000_000))
small_string = "".join(chr(random.randint(0, 255)) for _ in range(1000))

%timeit fast_serialize_string_value(string)
# 75.7 ms ± 178 μs per loop (mean ± std. dev. of 7 runs, 10 loops each)

%timeit slow_serialize_string_value(string)
# 954 ms ± 5.04 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

%timeit slow_serialize_string_value(string)
# 97 μs ± 1.22 μs per loop (mean ± std. dev. of 7 runs, 10,000 loops each)

%timeit fast_serialize_string_value(small_string)
# 6.8 μs ± 50.3 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)